arvello/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.0b2
chardet==5.2.0
Django==5.1.4
crispy_bootstrap5
django-crispy-forms==2.3
pillow==11.2.1
reportlab==4.3.1
six==1.16.0
soupsieve==2.6
sqlparse==0.5.3
tzdata==2024.1
django-clearcache==1.2.1
requests==2.32.3
python-barcode==0.15.1
pdf417gen==0.8.1
django-localflavor==4.0
django-bootstrap5==25.1
pandas==2.2.3
xhtml2pdf==0.2.17
openpyxl==3.1.5
django-simple-history==3.8.0
weasyprint==65.0
django-environ==0.12.0
python-decouple==3.8
#django-otp==1.5.1